Правки по ревью
