+++ packages/client/src/components/player/Player.tsx
@@ -1,30 +1,10 @@
-import React, { useState, useEffect } from 'react';
 import Button from 'react-bootstrap/Button';
+import useAudio from '../../hooks/useAudio';
 
 import './player.scss';
 
 type TPlayer = {
   url: string;
-};
-
-const useAudio = (url: string): [boolean, () => void] => {
-  const [audio] = useState(new Audio(url));
-  const [playing, setPlaying] = useState(false);
-
-  const toggle = () => setPlaying(!playing);
-
-  useEffect(() => {
-    playing ? audio.play() : audio.pause();
-  }, [playing]);
-
-  useEffect(() => {
-    audio.addEventListener('ended', () => setPlaying(false));
-    return () => {
-      audio.removeEventListener('ended', () => setPlaying(false));
-    };
-  }, []);
-
-  return [playing, toggle];
 };
 
 const Player = (props: TPlayer) => {
